fix(types): clear 1021 apps/web type errors and the live defects behind them
apps/web goes 1022 -> 1. packages/lib src/ reaches 0 (29 left, all in
scripts/). api 188 -> 44, worker 240 -> 9 and kb 161 -> 1 fall out of the
lib fixes without an edit. packages/ui, seed, billing, email, redis and
utils are swept for the first time and are all at 0.

Live defects fixed, each invisible except as a type error:

- Most built-in resources never fired bus events. unified-handler-mutations
  composed `${entityType}:${action}` unchecked, but only six entity types
  have per-type events, so part, quote, invoice, work_order, payment, tag
  and more published events no handler subscribes to: no timeline entry,
  no record rules, no dispatch or workflow trigger. Ticket status/assignee
  events also carried the wrong payload, writing recordId: undefined.
- postinstall-migrate.ts ran zero migrations while printing a success
  line on every deploy; it targeted a Prisma layout this repo dropped.
- SmartChunkingService called the static chunkText as an instance method,
  so dataset embedding threw on every document containing a blank line.
- 2FA OTP emails had no emailTypeSchema entry and dead-lettered.
- The OAuth consent flow was dead: no oidcClient on the client, and the
  handler read res.data.url where the endpoint returns redirectURI.
- The pnpm seed CLI was broken four ways (removed schema.Inbox, two
  non-existent columns, an onConflict target matching no constraint, and
  three of eight scenarios missing).
- Dataset segment pagination failed past page one and every segment
  optimistic update was a no-op; the datasets list was silently capped at
  20 with no pagination UI.
- Three lifecycle emails discarded every URL their job computed.
- zrevrange was entirely broken on Upstash.
- WebhookService.storeDelivery returned Ok(undefined) on a failed insert.

Also removes ~5300 lines of dead code: seed's abandoned relational-seeding
subsystem, lib's unused files/upload strategy layer, two Prisma-era mixin
factories, a MessageTypePicker whose options referenced non-existent enum
members, and assorted unreferenced hooks, selectors and scaffolding.

Full 12-project suite green (772 files, 9724 tests). Lint clean.

The one remaining error is AgentDomainConfig's invariance in TDomainState;
threading the generic properly cascades through AgentDefinition and the
pipeline helpers and is a separate refactor. Recorded in the handoff.

+++ apps/web/src/server/api/routers/dataset.ts
@@ -154,7 +154,10 @@ export const datasetRouter = createTRPCRouter({
         return {
           ...dataset,
           documentCount: Number(stats?.dc || 0),
-          totalSize: BigInt(Math.floor(stats?.ts || 0)),
+          // `Dataset.totalSize` is `bigint({ mode: 'number' })` — a number on both
+          // sides. Re-wrapping it as a real BigInt here made this branch's shape
+          // disagree with `DatasetEntity` for every consumer of `getById`.
+          totalSize: Math.floor(stats?.ts || 0),
         }
       }
       return dataset
